Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-images-to-the-canvas/2d.drawImage.broken-expected.txt

2d.drawImage.broken
Actual output:
Expected output:


FAIL Canvas test: 2d.drawImage.broken assert_approx_equals: Green channel of the pixel at (50, 25) expected 255 +/- 2 but got 0
